relocate types
