+++ apps/extension/src/state/tx-approval.ts
@@ -2,7 +2,7 @@ import { AuthorizeRequest } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumb
 import type { Jsonified, Stringified, TransactionClassification } from '@penumbra-zone/types';
 import { AllSlices, SliceCreator } from './index';
 import { MessageResponder } from '@penumbra-zone/types/src/internal-msg/shared';
-import { TxApproval } from '@penumbra-zone/types/src/internal-msg/tx-approval';
+import { TxApproval } from '@penumbra-zone/types/src/internal-msg/popup';
 import { TransactionView } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/transaction/v1/transaction_pb';
 import { Address } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/keys/v1/keys_pb';
 import { viewClient } from '../clients/extension-page';
@@ -14,6 +14,7 @@ import {
 } from '@penumbra-zone/types';
 import { JsonObject } from '@bufbuild/protobuf';
 import { ConnectError } from '@connectrpc/connect';
+
 export interface TxApprovalSlice {
   /**
    * Zustand doesn't like JsonValue, because the type is infinitely deep. And we
