+++ packages/router/src/ctx/approver.ts
@@ -1,8 +1,11 @@
 import { AuthorizeRequest } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/custody/v1/custody_pb';
 import { Jsonified } from '@penumbra-zone/types';
-import { sendPopupRequest, spawnDetachedPopup } from '@penumbra-zone/types/src/internal-msg/popup';
+import {
+  TxApproval,
+  sendPopupRequest,
+  spawnDetachedPopup,
+} from '@penumbra-zone/types/src/internal-msg/popup';
 import { TransactionView } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/transaction/v1/transaction_pb';
-import { TxApproval } from '@penumbra-zone/types/src/internal-msg/tx-approval';
 import { Code, ConnectError } from '@connectrpc/connect';
 
 export const getTxApproval = async (
